Split out realm portion of `RulesetStore`
