+++ osu.Game/Rulesets/RulesetStore.cs
@@ -7,34 +7,26 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using osu.Framework;
-using osu.Framework.Extensions.ObjectExtensions;
 using osu.Framework.Logging;
 using osu.Framework.Platform;
-using osu.Game.Database;
 
 #nullable enable
 
 namespace osu.Game.Rulesets
 {
-    public class RulesetStore : IDisposable, IRulesetStore
+    public abstract class RulesetStore : IDisposable, IRulesetStore
     {
-        private readonly RealmAccess realmAccess;
-
         private const string ruleset_library_prefix = @"osu.Game.Rulesets";
 
-        private readonly Dictionary<Assembly, Type> loadedAssemblies = new Dictionary<Assembly, Type>();
+        protected readonly Dictionary<Assembly, Type> LoadedAssemblies = new Dictionary<Assembly, Type>();
 
         /// <summary>
         /// All available rulesets.
         /// </summary>
-        public IEnumerable<RulesetInfo> AvailableRulesets => availableRulesets;
+        public abstract IEnumerable<RulesetInfo> AvailableRulesets { get; }
 
-        private readonly List<RulesetInfo> availableRulesets = new List<RulesetInfo>();
-
-        public RulesetStore(RealmAccess realm, Storage? storage = null)
+        protected RulesetStore(Storage? storage = null)
         {
-            realmAccess = realm;
-
             // On android in release configuration assemblies are loaded from the apk directly into memory.
             // We cannot read assemblies from cwd, so should check loaded assemblies instead.
             loadFromAppDomain();
@@ -53,8 +45,6 @@ namespace osu.Game.Rulesets
             var rulesetStorage = storage?.GetStorageForDirectory(@"rulesets");
             if (rulesetStorage != null)
                 loadUserRulesets(rulesetStorage);
-
-            addMissingRulesets();
         }
 
         /// <summary>
@@ -95,80 +85,7 @@ namespace osu.Game.Rulesets
             if (domainAssembly != null)
                 return domainAssembly;
 
-            return loadedAssemblies.Keys.FirstOrDefault(a => a.FullName == asm.FullName);
-        }
-
-        private void addMissingRulesets()
-        {
-            realmAccess.Write(realm =>
-            {
-                var rulesets = realm.All<RulesetInfo>();
-
-                List<Ruleset> instances = loadedAssemblies.Values
-                                                          .Select(r => Activator.CreateInstance(r) as Ruleset)
-                                                          .Where(r => r != null)
-                                                          .Select(r => r.AsNonNull())
-                                                          .ToList();
-
-                // add all legacy rulesets first to ensure they have exclusive choice of primary key.
-                foreach (var r in instances.Where(r => r is ILegacyRuleset))
-                {
-                    if (realm.All<RulesetInfo>().FirstOrDefault(rr => rr.OnlineID == r.RulesetInfo.OnlineID) == null)
-                        realm.Add(new RulesetInfo(r.RulesetInfo.ShortName, r.RulesetInfo.Name, r.RulesetInfo.InstantiationInfo, r.RulesetInfo.OnlineID));
-                }
-
-                // add any other rulesets which have assemblies present but are not yet in the database.
-                foreach (var r in instances.Where(r => !(r is ILegacyRuleset)))
-                {
-                    if (rulesets.FirstOrDefault(ri => ri.InstantiationInfo.Equals(r.RulesetInfo.InstantiationInfo, StringComparison.Ordinal)) == null)
-                    {
-                        var existingSameShortName = rulesets.FirstOrDefault(ri => ri.ShortName == r.RulesetInfo.ShortName);
-
-                        if (existingSameShortName != null)
-                        {
-                            // even if a matching InstantiationInfo was not found, there may be an existing ruleset with the same ShortName.
-                            // this generally means the user or ruleset provider has renamed their dll but the underlying ruleset is *likely* the same one.
-                            // in such cases, update the instantiation info of the existing entry to point to the new one.
-                            existingSameShortName.InstantiationInfo = r.RulesetInfo.InstantiationInfo;
-                        }
-                        else
-                            realm.Add(new RulesetInfo(r.RulesetInfo.ShortName, r.RulesetInfo.Name, r.RulesetInfo.InstantiationInfo, r.RulesetInfo.OnlineID));
-                    }
-                }
-
-                List<RulesetInfo> detachedRulesets = new List<RulesetInfo>();
-
-                // perform a consistency check and detach final rulesets from realm for cross-thread runtime usage.
-                foreach (var r in rulesets.OrderBy(r => r.OnlineID))
-                {
-                    try
-                    {
-                        var resolvedType = Type.GetType(r.InstantiationInfo)
-                                           ?? throw new RulesetLoadException(@"Type could not be resolved");
-
-                        var instanceInfo = (Activator.CreateInstance(resolvedType) as Ruleset)?.RulesetInfo
-                                           ?? throw new RulesetLoadException(@"Instantiation failure");
-
-                        // If a ruleset isn't up-to-date with the API, it could cause a crash at an arbitrary point of execution.
-                        // To eagerly handle cases of missing implementations, enumerate all types here and mark as non-available on throw.
-                        resolvedType.Assembly.GetTypes();
-
-                        r.Name = instanceInfo.Name;
-                        r.ShortName = instanceInfo.ShortName;
-                        r.InstantiationInfo = instanceInfo.InstantiationInfo;
-                        r.Available = true;
-
-                        detachedRulesets.Add(r.Clone());
-                    }
-                    catch (Exception ex)
-                    {
-                        r.Available = false;
-                        Logger.Log($"Could not load ruleset {r}: {ex.Message}");
-                    }
-                }
-
-                availableRulesets.AddRange(detachedRulesets.OrderBy(r => r));
-            });
+            return LoadedAssemblies.Keys.FirstOrDefault(a => a.FullName == asm.FullName);
         }
 
         private void loadFromAppDomain()
@@ -214,7 +131,7 @@ namespace osu.Game.Rulesets
         {
             string? filename = Path.GetFileNameWithoutExtension(file);
 
-            if (loadedAssemblies.Values.Any(t => Path.GetFileNameWithoutExtension(t.Assembly.Location) == filename))
+            if (LoadedAssemblies.Values.Any(t => Path.GetFileNameWithoutExtension(t.Assembly.Location) == filename))
                 return;
 
             try
@@ -229,17 +146,17 @@ namespace osu.Game.Rulesets
 
         private void addRuleset(Assembly assembly)
         {
-            if (loadedAssemblies.ContainsKey(assembly))
+            if (LoadedAssemblies.ContainsKey(assembly))
                 return;
 
             // the same assembly may be loaded twice in the same AppDomain (currently a thing in certain Rider versions https://youtrack.jetbrains.com/issue/RIDER-48799).
             // as a failsafe, also compare by FullName.
-            if (loadedAssemblies.Any(a => a.Key.FullName == assembly.FullName))
+            if (LoadedAssemblies.Any(a => a.Key.FullName == assembly.FullName))
                 return;
 
             try
             {
-                loadedAssemblies[assembly] = assembly.GetTypes().First(t => t.IsPublic && t.IsSubclassOf(typeof(Ruleset)));
+                LoadedAssemblies[assembly] = assembly.GetTypes().First(t => t.IsPublic && t.IsSubclassOf(typeof(Ruleset)));
             }
             catch (Exception e)
             {
